misc(b-modal): remove temporary CSS for vertically centered modal
